mask of type int

+++ mtg/ml/models.py
@@ -624,13 +624,21 @@ class DeckBuilder(tf.Module):
         if self.card_embeddings is not None:
             pool_embs = pools[:, :, :, None] * self.card_embeddings[None, None, :, :]
             deck_embs = decks[:, :, :, None] * self.card_embeddings[None, None, :, :]
-            deck_mask = tf.where(tf.cast(decks, dtype=tf.int32) > 0, 0, 1)
-            pool_mask = tf.where(tf.cast(pools, dtype=tf.int32) > 0, 0, 1)
+            deck_mask = tf.where(decks > 0, 0, 1)
+            pool_mask = tf.where(pools > 0, 0, 1)
             deck_att, _ = self.deck_attention(
-                deck_embs, deck_embs, deck_embs, mask=deck_mask, training=training
+                deck_embs,
+                deck_embs,
+                deck_embs,
+                mask=tf.cast(deck_mask, dtype=tf.int32),
+                training=training,
             )
             pool_att, _ = self.pool_attention(
-                pool_embs, pool_embs, pool_embs, mask=pool_mask, training=training
+                pool_embs,
+                pool_embs,
+                pool_embs,
+                mask=tf.cast(pool_mask, dtype=tf.int32),
+                training=training,
             )
             self.latent_rep_pool = tf.reduce_sum(pool_att, axis=2)
             self.latent_rep_deck = tf.reduce_sum(deck_att, axis=2)
